backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-dotenv==1.0.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0

# RAG & AI
# IMPORTANT: Install PyTorch CPU-only separately first (much smaller ~200MB vs 2GB+):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
langchain==0.0.350
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
pypdf==3.17.1
openai>=1.12.0

# File Handling
aiofiles==23.2.1
python-magic==0.4.27

# CORS is built into FastAPI

